update a process of tasks' computation
Now each task specifies a computation model to being used
to compute the task. Using this specification a proper computer
(tasks' executor) gets chosen and computation happens on the level
of the computer, so the task itself doesn't define how it should
be computed, it deligates this to the specific computer.

Now only one computer has been implemented: raw_src_computer,
which is responsible for the computation of calc_chunk tasks
(raw sources + data as an array of bytes).

Also classes to handle dynamic modules, compiling such dynamic
modules and class managing a non-open file (regular_file) have
been implemented.